requirements.txt
requests
airtable-python-wrapper
flask